requirements.txt
discord.py
python-dotenv
requests
aiosqlite
asyncio
flask
gunicorn
psycopg2-binary
pyotp
setuptools